// src/webapp/classicladder/src/generated/classicladder_client.ts
export enum LadderState {
  LOADING = 0,
  STOP = 1,
  RUN = 2,
}

export enum SectionLanguage {
  LADDER = 0,
  SEQUENTIAL = 1,
}

export enum TimerIECMode {
  TON = 0,
  TOF = 1,
  TP = 2,
}

export enum ModbusReqType {
  COIL_READ = 0,
  COIL_WRITE = 1,
  INPUT_READ = 2,
  HOLDING_READ = 3,
  REGISTER_WRITE = 4,
  DIAGNOSTIC = 5,
}

// --- Types ---

export interface Element {
  type: number;
  connectedWithTop: number;
  varType: number;
  varNum: number;
}

export interface Rung {
  used: boolean;
  prevRung: number;
  nextRung: number;
  label: string;
  comment: string;
  elements: Element[];
}

export interface Section {
  used: boolean;
  name: string;
  language: SectionLanguage;
  subRoutineNumber: number;
  firstRung: number;
  lastRung: number;
  sequentialPage: number;
}

export interface Symbol {
  varName: string;
  symbol: string;
  comment: string;
}

export interface ArithmExpr {
  expr: string;
}

export interface TimerIEC {
  preset: number;
  value: number;
  base: number;
  mode: TimerIECMode;
  input: boolean;
  output: boolean;
}

export interface Timer {
  preset: number;
  value: number;
  base: number;
  inputEnable: boolean;
  inputControl: boolean;
  outputDone: boolean;
  outputRunning: boolean;
}

export interface Monostable {
  preset: number;
  value: number;
  base: number;
  input: boolean;
  outputRunning: boolean;
}

export interface Counter {
  preset: number;
  value: number;
  inputReset: boolean;
  inputPreset: boolean;
  inputCountUp: boolean;
  inputCountDown: boolean;
  outputDone: boolean;
  outputEmpty: boolean;
  outputFull: boolean;
}

export interface SizeInfo {
  nbrRungs: number;
  nbrBits: number;
  nbrWords: number;
  nbrTimers: number;
  nbrMonostables: number;
  nbrCounters: number;
  nbrTimersIec: number;
  nbrPhysInputs: number;
  nbrPhysOutputs: number;
  nbrArithmExpr: number;
  nbrSections: number;
  nbrSymbols: number;
  nbrS32in: number;
  nbrS32out: number;
  nbrFloatIn: number;
  nbrFloatOut: number;
  nbrErrorBits: number;
}

export interface Status {
  state: LadderState;
  scanTimeUs: number;
  periodicRefreshMs: number;
  sizes: SizeInfo;
  projectFile: string;
}

export interface Program {
  sizes: SizeInfo;
  rungs: Rung[];
  sections: Section[];
  symbols: Symbol[];
  arithmExprs: ArithmExpr[];
  timersIec: TimerIEC[];
  timers: Timer[];
  monostables: Monostable[];
  counters: Counter[];
}